requirements.txt
camel-ai
mistralai
streamlit
python-dotenv
selenium 
duckduckgo-search